+++ packages/delir-core/src/Values/index.ts
@@ -4,4 +4,4 @@ export {default as Size2D} from './size-2d'
 export {default as Size3D} from './size-3d'
 export {default as ColorRGB} from './color-rgb'
 export {default as ColorRGBA} from './color-rgba'
-export {default as Expression} from './expression'
+export {default as Expression} from './Expression'
